+++ apps/docs/src/components/actions/Menu.stories.tsx
@@ -1,91 +1,199 @@
-import * as React from "react";
-import { Meta, Story } from "@storybook/react";
+import React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
 import { Menu } from "@react-elf/ui";
+import { MdDescription } from "react-icons/md";
 
-export default {
+const meta = {
   component: Menu,
-  title: "Components/Actions/Menu",
-  argTypes: {
-    type: {
-      name: "type",
-      description: "type of menu",
-      type: { name: "string", required: true },
-      control: {
-        type: "select",
-        options: ["menu", "dropdown", "contextmenu"],
-      },
-    },
-    hover: {
-      name: "hover",
-      description: "hover",
-      type: { name: "boolean", required: false },
-      control: {
-        type: "boolean",
-      },
-    },
-    compact: {
-      name: "compact",
-      description: "compact",
-      type: { name: "boolean", required: false },
-      control: {
-        type: "boolean",
-      },
-    },
-  },
-  args: {
-    type: "menu",
-    hover: false,
-  },
-  parameters: {
-    componentSubtitle:
-      "Buttons should always have a label, unless they are only using an icon that is universally understood and accessible. ",
-    docs: {
-      description: {
-        component: `
-They can have an optional icon, but it should not be used for decoration. Use an icon only when necessary and when it has a strong association with the label text.
+  tags: ["autodocs"],
+} satisfies Meta<typeof Menu>;
 
-The label can be hidden to create an icon-only button. If the label is hidden, an icon is required, and the label will appear in a tooltip.
-        `,
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    items: [
+      { type: "item", title: "Document Setup" },
+      { type: "item", title: "Settings" },
+      { type: "item", title: "Copy" },
+      {
+        type: "section",
+        title: "Group",
       },
+      {
+        type: "item",
+        title: "Languages",
+        hover: true,
+        items: [
+          { type: "item", title: "English", selectable: true },
+          {
+            type: "item",
+            title: "German",
+            selected: true,
+            selectable: true,
+          },
+          { type: "item", title: "French", selectable: true },
+          { type: "item", title: "Spanish", selectable: true },
+          { type: "item", title: "Italian", selectable: true },
+          "-",
+          { type: "item", title: "Portuguese", selectable: true },
+          { type: "item", title: "Russian", selectable: true },
+          { type: "item", title: "Japanese", selectable: true },
+          { type: "item", title: "Chinese", selectable: true },
+        ],
+      },
+    ],
+    style: {
+      position: "relative",
     },
   },
-} as Meta;
-export const DefaultMenu: Story = (args) => {
-  const { type, hover, compact } = args;
-  return (
-    <div style={{ display: "flex", gap: 10, flexDirection: "column" }}>
-      <div style={{ display: "flex", gap: 10, padding: 100 }}>
-        <Menu
-          type={type}
-          compact={compact}
-          items={[
-            {
-              type: "item",
-              title: "Item 1",
-            },
-            {
-              type: "divider",
-              dashed: true,
-            },
-            { type: "section", title: "Section title" },
-            {
-              type: "item",
-              title: "Item 2",
-            },
-            {
-              type: "item",
-              title: "Item 3",
-              hover: hover,
-              items: [
-                {
-                  type: "item",
-                  title: "aaaa",
-                },
-              ],
-            },
-          ]}
-        />
-      </div>
-    </div>
-  );
+};
+
+export const PopoverContainer: Story = {
+  args: {
+    items: [
+      { type: "item", title: "Document Setup" },
+      { type: "item", title: "Settings" },
+      { type: "item", title: "Copy" },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Icon: Story = {
+  args: {
+    items: [
+      { type: "item", title: "Document Setup", icon: <MdDescription /> },
+      { type: "item", title: "Settings" },
+      { type: "item", title: "Copy" },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Description: Story = {
+  args: {
+    items: [
+      {
+        type: "item",
+        title: "Quick export",
+        description: "Share a low-res snapshot",
+      },
+      { type: "item", title: "Settings" },
+      {
+        type: "item",
+        title: "Open a copy",
+        description: "Open in iPad",
+      },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Value: Story = {
+  args: {
+    items: [
+      {
+        type: "item",
+        title: "Quick export",
+        shortcut: "⌘ + Ctrl+E",
+        hover: true,
+        description: "Share a low-res snapshot",
+        items: [{ type: "item", title: "Document Setup" }],
+      },
+      { type: "item", title: "Settings" },
+      {
+        type: "item",
+        title: "Open a copy",
+        description: "Open in iPad",
+      },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    items: [
+      {
+        type: "item",
+        title: "Quick export",
+        shortcut: "⌘ + Ctrl+E",
+        disabled: true,
+        description: "Share a low-res snapshot",
+        items: [{ type: "item", title: "Document Setup" }],
+      },
+      { type: "item", title: "Settings" },
+      {
+        type: "item",
+        title: "Open a copy",
+        disabled: true,
+        description: "Open in iPad",
+      },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Section: Story = {
+  args: {
+    items: [
+      {
+        type: "section",
+        title: "Section 1",
+      },
+      {
+        type: "item",
+        title: "Quick export",
+        shortcut: "⌘ + Ctrl+E",
+        description: "Share a low-res snapshot",
+        items: [{ type: "item", title: "Document Setup" }],
+      },
+      { type: "item", title: "Settings" },
+      "-",
+      {
+        type: "item",
+        title: "Open a copy",
+        description: "Open in iPad",
+      },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
+};
+
+export const Variant: Story = {
+  args: {
+    variant: "dark",
+    items: [
+      {
+        type: "item",
+        title: "Quick export",
+        shortcut: "⌘ + Ctrl+E",
+        description: "Share a low-res snapshot",
+        items: [{ type: "item", title: "Document Setup" }],
+      },
+      { type: "item", title: "Settings" },
+      {
+        type: "item",
+        title: "Open a copy",
+        description: "Open in iPad",
+      },
+    ],
+    style: {
+      position: "relative",
+    },
+  },
 };