+++ packages/@react-elf-types/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @react-elf-types/shared
 
+## 0.0.65
+
+### Patch Changes
+
+- add Menu stories
+
 ## 0.0.60
 
 ### Patch Changes