+++ packages/@react-elf-types/toolbar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @react-elf-types/toolbar
 
+## 0.0.65
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.65
+  - @react-elf-types/tools@0.0.65
+
 ## 0.0.64
 
 ### Patch Changes