+++ packages/@react-elf-types/tools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf-types/tools
 
+## 0.0.65
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.65
+  - @react-elf-types/menu@0.0.65
+  - @react-elf-types/tooltip@0.0.65
+
 ## 0.0.60
 
 ### Patch Changes